refactored game domain to reflect best practice domain driven design package structure
