--- backend/src/main/java/com/gleam/backend/dto/LoginDto.java
@@ -1,4 +1,0 @@
-package com.gleam.backend.dto;
-
-public record LoginDto(String username, String password) {
-}
