--- backend/src/main/java/com/gleam/backend/dto/ReviewDto.java
@@ -1,4 +1,0 @@
-package com.gleam.backend.dto;
-
-public record ReviewDto(String title, String comment) {
-}
