+++ backend/src/main/java/com/gleam/backend/game/config/GleamCorsConfiguration.java
@@ -1,4 +1,4 @@
-package com.gleam.backend.config;
+package com.gleam.backend.game.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
