+++ backend/src/main/java/com/gleam/backend/game/infrastructure/GameRepository.java
@@ -1,6 +1,6 @@
-package com.gleam.backend.repository;
+package com.gleam.backend.game.infrastructure;
 
-import com.gleam.backend.model.Game;
+import com.gleam.backend.game.domain.Game;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
