+++ backend/src/main/java/com/gleam/backend/game/ui/dto/GameDto.java
@@ -1,4 +1,4 @@
-package com.gleam.backend.dto;
+package com.gleam.backend.game.ui.dto;
 
 import com.fasterxml.jackson.annotation.JsonFormat;
 import org.springframework.data.annotation.Id;
