--- backend/src/main/java/com/gleam/backend/mapper/RatingMapper.java
@@ -1,11 +1,0 @@
-package com.gleam.backend.mapper;
-
-import com.gleam.backend.dto.ReviewDto;
-import com.gleam.backend.model.Review;
-import org.mapstruct.Mapper;
-
-@Mapper(componentModel = "spring")
-public interface RatingMapper {
-
-    Review toEntity(ReviewDto reviewDto);
-}
